+++ PRIO/Models/Wells/Well.cs
@@ -17,9 +17,9 @@ namespace PRIO.Models.Wells
         public string? CategoryOperator { get; set; }
         public bool? StatusOperator { get; set; }
         public string? Type { get; set; }
-        public double? WaterDepth { get; set; }
-        public double? TopOfPerforated { get; set; }
-        public double? BaseOfPerforated { get; set; }
+        public decimal? WaterDepth { get; set; }
+        public decimal? TopOfPerforated { get; set; }
+        public decimal? BaseOfPerforated { get; set; }
         public string? ArtificialLift { get; set; }
         public string? Latitude4C { get; set; }
         public string? Longitude4C { get; set; }
